+++ features/step_definitions/rails_app.rb
@@ -1,15 +1,15 @@
 def setup_rails_app
-  return if File.directory?("test/rails_app")
+  return if File.directory?('test/rails_app')
 
-  unless system "bundle exec rails new test/rails_app -m test/rails_template.rb && cd ./test/rails_app && RAILS_ENV=test rake db:migrate"
-    system("rm -fr test/rails_app")
-    raise "Failed to generate test/rails_app"
+  unless system 'bundle exec rails new test/rails_app -m test/rails_template.rb && cd ./test/rails_app && RAILS_ENV=test rake db:migrate'
+    system('rm -fr test/rails_app')
+    fail 'Failed to generate test/rails_app'
   end
 end
 
 Given /^I have a rails app using 'active_sanity'$/ do
-  Dir["./test/rails_app/db/migrate/*create_invalid_records.rb"].each do |migration|
-    raise unless system("rm #{migration}")
+  Dir['./test/rails_app/db/migrate/*create_invalid_records.rb'].each do |migration|
+    fail unless system("rm #{migration}")
   end
 
   setup_rails_app
@@ -23,7 +23,7 @@ end
 Given /^I have a rails app using 'active_sanity' with db storage$/ do
   setup_rails_app
 
-  raise unless system("cd ./test/rails_app && rails generate active_sanity && RAILS_ENV=test rake db:migrate")
+  fail unless system('cd ./test/rails_app && rails generate active_sanity && RAILS_ENV=test rake db:migrate')
 
   require './test/rails_app/config/environment'
 
@@ -33,16 +33,16 @@ Given /^I have a rails app using 'active_sanity' with db storage$/ do
 end
 
 Given /^the database contains a few valid records$/ do
-  Author.create!(:first_name => "Greg", :last_name => "Bell", :username => "gregbell")
-  Publisher.create!(:first_name => "Sam",  :last_name => "Vincent", :username => "samvincent")
-  Category.create!(:name => "Uncategorized")
-  Post.create!(:author => Author.first, :category => Category.first,
-    :title => "How ActiveAdmin changed the world", :body => "Lot of love.",
-    :published_at => 4.years.from_now)
+  Author.create!(first_name: 'Greg', last_name: 'Bell', username: 'gregbell')
+  Publisher.create!(first_name: 'Sam',  last_name: 'Vincent', username: 'samvincent')
+  Category.create!(name: 'Uncategorized')
+  Post.create!(author: Author.first, category: Category.first,
+    title: 'How ActiveAdmin changed the world', body: 'Lot of love.',
+    published_at: 4.years.from_now)
 end
 
 Given /^the first author's username is empty and the first post category_id is nil$/ do
-  Author.first.update_attribute(:username, "")
+  Author.first.update_attribute(:username, '')
   Post.first.update_attribute(:category_id, nil)
 end
 
@@ -58,12 +58,10 @@ Given /^the first post title is empty$/ do
   Post.first.update_attribute('title', '')
 end
 
-
 When /^I run "([^"]*)"$/ do |command|
   puts @output = `cd ./test/rails_app && export RAILS_ENV=test && bundle exec #{command} --trace; echo "RETURN:$?"`
-  raise unless @output['RETURN:0']
+  fail unless @output['RETURN:0']
 end
-
 
 Then /^I should see the following invalid records:$/ do |table|
   table.raw.each do |model, id, errors|
@@ -85,7 +83,7 @@ end
 
 Then /^the table "([^"]*)" should contain:$/ do |_, table|
   table.raw.each do |model, id, errors|
-    invalid_record = InvalidRecord.where(:record_type => model, :record_id => id).first
+    invalid_record = InvalidRecord.where(record_type: model, record_id: id).first
     invalid_record.should be_an_instance_of(InvalidRecord)
     errors = eval(errors)
     errors.each do |k, v|
@@ -95,6 +93,5 @@ Then /^the table "([^"]*)" should contain:$/ do |_, table|
 end
 
 Then /^the table "([^"]*)" should not contain errors for "([^"]*)" "([^"]*)"$/ do |_, model, id|
-    InvalidRecord.where(:record_type => model, :record_id => id).first.should be_nil
+  InvalidRecord.where(record_type: model, record_id: id).first.should be_nil
 end
-