+++ lib/active_sanity/invalid_record.rb
@@ -1,5 +1,5 @@
 class InvalidRecord < ActiveRecord::Base
-  belongs_to :record, :polymorphic => true
+  belongs_to :record, polymorphic: true
   serialize :validation_errors
 
   validates_presence_of :record, :validation_errors