+++ lib/active_sanity/tasks.rb
@@ -1,7 +1,6 @@
 namespace :db do
-  desc "Check records sanity"
-  task :check_sanity => :environment do
+  desc 'Check records sanity'
+  task check_sanity: :environment do
     ActiveSanity::Checker.check!
   end
 end
-